Make PdfObject#getType return byte instead of int

+++ kernel/src/main/java/com/itextpdf/kernel/pdf/PdfLiteral.java
@@ -27,7 +27,7 @@ public class PdfLiteral extends PdfPrimitiveObject {
     }
 
     @Override
-    public int getType() {
+    public byte getType() {
         return Literal;
     }
